+++ erpnext/controllers/stock_controller.py
@@ -563,7 +563,7 @@ class StockController(AccountsController):
 
 @frappe.whitelist()
 def show_accounting_ledger_preview(company: str, doctype: str, docname: str):
-	from erpnext.stock.services.ledger_preview import get_accounting_ledger_preview
+	from erpnext.controllers.ledger_preview import get_accounting_ledger_preview
 
 	filters = frappe._dict(company=company, include_dimensions=1)
 	doc = frappe.get_lazy_doc(doctype, docname)
@@ -578,7 +578,7 @@ def show_accounting_ledger_preview(company: str, doctype: str, docname: str):
 
 @frappe.whitelist()
 def show_stock_ledger_preview(company: str, doctype: str, docname: str):
-	from erpnext.stock.services.ledger_preview import get_stock_ledger_preview
+	from erpnext.controllers.ledger_preview import get_stock_ledger_preview
 
 	filters = frappe._dict(company=company)
 	doc = frappe.get_lazy_doc(doctype, docname)
